Render most of page with React
I think this mostly works.  There may be some racy behavior on load that needs
to be sorted out between the backend rendering JSON and the frontend
Page.getInitialState().  But otherwise this is probably okay.

Most components are pure functions on props, so we could add PureRenderMixin for
better performance (if needed).