Real Django data. Real Django updates!
You can modify an answer, notes field, or tags field on a puzzle and
it will update the server accordingly.

TODO: make a unidirectional data flow with flux so that you don't
have to wait for the next ping of the server to see your update. Or
make it so components can trigger server pings, I guess.